backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
pdf2image==1.16.3
pillow==10.0.0
pytesseract==0.3.10
gunicorn==20.1.0
